modify documentation, move fasta_list to sklearn_compatiblity tests
